+++ cli/build.gradle
@@ -1,0 +1,16 @@
+dependencies {
+    implementation project(':core')
+}
+
+sourceSets {
+    main {
+        resources {
+            srcDirs += [
+                    project(':core').sourceSets.main.resources
+            ]
+        }
+    }
+}
+
+bootJar { enabled = true }
+jar { enabled = false }
