+++ core/src/main/java/bear/replacer/core/excel/model/ExcelLoadFailedException.java
@@ -1,0 +1,8 @@
+package bear.replacer.core.excel.model;
+
+public class ExcelLoadFailedException extends RuntimeException {
+
+    public ExcelLoadFailedException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
