+++ core/src/main/java/bear/replacer/core/excel/model/ExcelWriteFailedException.java
@@ -1,0 +1,8 @@
+package bear.replacer.core.excel.model;
+
+public class ExcelWriteFailedException extends RuntimeException {
+
+    public ExcelWriteFailedException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
