+++ core/src/main/java/bear/replacer/core/rule/RuleLoader.java
@@ -1,0 +1,10 @@
+package bear.replacer.core.rule;
+
+import bear.replacer.core.rule.model.Rule;
+
+import java.util.List;
+
+public interface RuleLoader {
+
+    List<Rule> load();
+}
