+++ core/src/main/resources/rule/rule.csv
