+++ core/src/test/resources/rule.csv
