+++ gradle.properties
@@ -1,1 +1,1 @@
-org.gradle.jvmargs=-Dfile.encoding=UTF-8
+org.gradle.jvmargs=-Dfile.encoding=UTF-8 -Dconsole.encoding=UTF-8
