+++ settings.gradle
@@ -1,1 +1,2 @@
 rootProject.name = 'excel-word-replacer'
+include 'core', 'cli'
