--- src/main/java/bear/excel/replacer/ExcelConverterApplication.java
@@ -1,59 +1,0 @@
-package bear.excel.replacer;
-
-import bear.excel.replacer.excel.ExcelLoader;
-import bear.excel.replacer.excel.ExcelReplacer;
-import bear.excel.replacer.excel.ExcelWriter;
-import bear.excel.replacer.rule.CsvRuleLoader;
-import bear.excel.replacer.rule.RuleLoader;
-import bear.excel.replacer.rule.model.Rule;
-import org.apache.poi.ss.usermodel.Workbook;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.boot.CommandLineRunner;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-import java.io.IOException;
-import java.util.List;
-
-@SpringBootApplication
-public class ExcelConverterApplication implements CommandLineRunner {
-
-    private static final Logger LOG = LoggerFactory.getLogger(ExcelConverterApplication.class);
-
-    public static void main(String[] args) {
-        SpringApplication.run(ExcelConverterApplication.class, args);
-    }
-
-    @Override
-    public void run(String... args) {
-        if (args.length != 2) {
-            System.out.println(args[0]);
-            LOG.info("1. 사용방법: replacer <inputFilePath> <outputFilePath>");
-            LOG.info("2. 예시: replacer ./test-excel.xlsx ./test-excel2.xlsx");
-            return;
-        }
-
-        final String input = args[0];
-        final String output = args[1];
-
-        LOG.info("대상 파일: {}", input);
-        final ExcelLoader excelLoader = new ExcelLoader(input);
-
-        try (final Workbook workbook = excelLoader.load()) {
-            final RuleLoader ruleLoader = new CsvRuleLoader(ClassLoader.getSystemResourceAsStream("rule\\rule.csv"));
-            final List<Rule> rules = ruleLoader.load();
-            LOG.info("대체할 문자열: {}", rules);
-
-            final ExcelReplacer replacer = new ExcelReplacer(workbook, rules);
-            final Workbook replaced = replacer.replace();
-
-            LOG.info("Replacing workbook to: {}", output);
-
-            final ExcelWriter excelWriter = new ExcelWriter();
-            excelWriter.write(replaced, output);
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-    }
-}
