--- src/main/java/bear/excel/replacer/excel/ExcelLoader.java
@@ -1,22 +1,0 @@
-package bear.excel.replacer.excel;
-
-import bear.excel.replacer.excel.model.ExcelLoadFailedException;
-import lombok.RequiredArgsConstructor;
-import org.apache.poi.ss.usermodel.Workbook;
-import org.apache.poi.xssf.usermodel.XSSFWorkbook;
-
-import java.io.FileInputStream;
-
-@RequiredArgsConstructor
-public class ExcelLoader {
-
-    private final String filePath;
-
-    public Workbook load() {
-        try (FileInputStream fis = new FileInputStream(filePath)) {
-            return new XSSFWorkbook(fis);
-        } catch (Exception e) {
-            throw new ExcelLoadFailedException("Cannot load excel file: " + filePath, e);
-        }
-    }
-}
