--- src/main/java/bear/excel/replacer/excel/model/ExcelLoadFailedException.java
@@ -1,8 +1,0 @@
-package bear.excel.replacer.excel.model;
-
-public class ExcelLoadFailedException extends RuntimeException {
-
-    public ExcelLoadFailedException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
