--- src/main/java/bear/excel/replacer/excel/model/ExcelWriteFailedException.java
@@ -1,8 +1,0 @@
-package bear.excel.replacer.excel.model;
-
-public class ExcelWriteFailedException extends RuntimeException {
-
-    public ExcelWriteFailedException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
