--- src/main/java/bear/excel/replacer/rule/RuleLoader.java
@@ -1,10 +1,0 @@
-package bear.excel.replacer.rule;
-
-import bear.excel.replacer.rule.model.Rule;
-
-import java.util.List;
-
-public interface RuleLoader {
-
-    List<Rule> load();
-}
