--- src/main/java/bear/excel/replacer/rule/model/Rule.java
@@ -1,17 +1,0 @@
-package bear.excel.replacer.rule.model;
-
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import lombok.ToString;
-
-@Getter
-@ToString
-@EqualsAndHashCode
-@RequiredArgsConstructor
-public class Rule {
-
-    private final String word;
-    private final String replaceWord;
-
-}
