--- src/main/resources/application.yml
@@ -1,4 +1,0 @@
-spring:
-  main:
-    banner-mode: off
-    web-application-type: none
